architecture: added interface for engines, which include the agent class, the environment class, and the trainer class
The task for an agent is to make decisions. For instance an agent can be a classifier, a feature extractor or a feature transformer. Stacking multiple agents together is an additional feature in the future.

An environment is where the agent makes decision, where the dataset or a simulator is located

A trainer is the class where the interaction between the agent and the environment is defined.